requirements.txt
streamlit
pandas
gspread
oauth2client
fpdf
Pillow